Phase 3: split pyproject — root keeps dev tooling, server gets its own
- New server/python/pyproject.toml: name=entdb-server, AGPL-3.0-only,
  scripts entdb/entdb-restore/entdb-schema → entdb_server.* entry points,
  base deps + optional extras (pubsub, servicebus, eventhubs, gcs,
  azure-storage, metrics, tracing).
- sdk/python/pyproject.toml unchanged (already moved in Phase 1; package
  name 'entdb-sdk' on PyPI preserved).
- Root pyproject: drops [project] (workspace-only root), keeps ruff +
  mypy + pytest + coverage. Adds pythonpath=[server/python, sdk/python]
  so pytest discovers the moved packages, and testpaths=tests/python.
  Updates known-first-party / mypy module overrides to the new
  entdb_server / entdb_sdk top-level names.

Test fixups so 2447 tests still pass under the new layout:
- bulk-rewrite sdk.entdb_sdk → entdb_sdk in tests + examples/demo.py
  (missed in Phase 2 — the package is imported as the top-level name,
  not as sdk.entdb_sdk).
- tests._test_schemas → tests.python._test_schemas (test schemas moved
  with the rest of tests/).
- test_compliance_docs / test_codegen_options_pb2: parents[2]→parents[3]
  (one extra path level for tests/python/).
- test_proto_options + test_lint: update hard-coded paths to point at
  sdk/python and tests/python.

Verified: python -m pytest tests/python/unit/ tests/python/integration/
=> 2447 passed (matches pre-restructure baseline).
