+++ pyproject.toml
@@ -1,103 +1,10 @@
-[build-system]
-requires = ["setuptools>=68.0", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "entdb"
-version = "1.7.0"
-description = "Tenant-sharded graph database with nodes and edges"
-readme = "README.md"
-license = "AGPL-3.0-only"
-requires-python = ">=3.10"
-authors = [
-    {name = "EntDB Team", email = "team@entdb.io"}
-]
-classifiers = [
-    "Development Status :: 4 - Beta",
-    "Intended Audience :: Developers",
-    "Operating System :: OS Independent",
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
-    "Topic :: Database",
-    "Topic :: Software Development :: Libraries :: Python Modules",
-    "Typing :: Typed",
-]
-keywords = ["database", "graph", "tenant", "sharding", "event-sourcing"]
-
-dependencies = []
-
-[project.optional-dependencies]
-server = [
-    "aiokafka>=0.10.0",
-    "aiobotocore>=2.11.0",
-    "aiohttp>=3.9.0",
-    "grpcio>=1.60.0",
-    "grpcio-health-checking>=1.60.0",
-    "PyJWT>=2.8.0",
-    "cryptography>=41.0.0",
-    "PyYAML>=6.0.0",
-]
-pubsub = [
-    "google-cloud-pubsub>=2.18.0",
-]
-servicebus = [
-    "azure-servicebus>=7.11.0",
-]
-eventhubs = [
-    "azure-eventhub>=5.11.0",
-]
-azure-storage = [
-    "azure-storage-blob>=12.19.0",
-]
-gcs = [
-    "google-cloud-storage>=2.14.0",
-]
-metrics = [
-    "prometheus_client>=0.21.0",
-]
-tracing = [
-    "opentelemetry-api>=1.20.0",
-    "opentelemetry-sdk>=1.20.0",
-    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
-]
-sdk = [
-    "grpcio>=1.60.0",
-    "protobuf>=4.25.0",
-]
-dev = [
-    "entdb[server,sdk]",
-    "pytest>=8.0.0",
-    "pytest-asyncio>=0.23.0",
-    "pytest-cov>=4.1.0",
-    "pytest-timeout>=2.2.0",
-    "ruff>=0.15.0",
-    "mypy>=1.8.0",
-    "httpx>=0.26.0",
-    "grpcio-tools>=1.60.0",
-]
-all = [
-    "entdb[server,sdk,metrics,tracing,pubsub,servicebus,eventhubs,azure-storage,gcs,dev]",
-]
-
-[project.scripts]
-entdb = "dbaas.entdb_server.main:main"
-entdb-restore = "dbaas.entdb_server.tools.restore:main"
-entdb-schema = "dbaas.entdb_server.tools.schema_cli:main"
-
-[project.urls]
-Homepage = "https://github.com/elloloop/tenant-shard-db"
-Documentation = "https://github.com/elloloop/tenant-shard-db#readme"
-Repository = "https://github.com/elloloop/tenant-shard-db"
-Issues = "https://github.com/elloloop/tenant-shard-db/issues"
-
-[tool.setuptools.packages.find]
-where = ["."]
-include = ["dbaas*", "sdk*"]
-
-[tool.setuptools.package-data]
-"*" = ["py.typed", "*.proto"]
+# EntDB monorepo — root pyproject.
+#
+# This file is NOT an installable package. It hosts shared dev tooling
+# config (ruff, mypy, pytest, coverage) for the polyglot workspace.
+# Installable Python packages live under:
+#   - server/python/pyproject.toml  (entdb-server, AGPL-3.0-only)
+#   - sdk/python/pyproject.toml     (entdb-sdk, MIT)
 
 # =============================================================================
 # Ruff (linter)
@@ -140,7 +47,7 @@ ignore = [
 ]
 
 [tool.ruff.lint.isort]
-known-first-party = ["dbaas", "sdk", "entdb_sdk", "entdb_server"]
+known-first-party = ["entdb_sdk", "entdb_server"]
 
 # =============================================================================
 # Mypy (type checker)
@@ -155,6 +62,7 @@ check_untyped_defs = true
 ignore_missing_imports = true
 show_error_codes = true
 strict_optional = true
+mypy_path = "server/python:sdk/python"
 exclude = [
     "_generated/",
     "api/generated/",
@@ -173,27 +81,27 @@ ignore_missing_imports = true
 module = [
     "entdb_sdk._generated",
     "entdb_sdk._generated.*",
-    "dbaas.entdb_server.api.generated",
-    "dbaas.entdb_server.api.generated.*",
+    "entdb_server.api.generated",
+    "entdb_server.api.generated.*",
 ]
 follow_imports = "skip"
 ignore_errors = true
 
 [[tool.mypy.overrides]]
 module = [
-    "dbaas.entdb_server.wal.kinesis",
-    "dbaas.entdb_server.wal.kafka",
-    "dbaas.entdb_server.wal.pubsub",
-    "dbaas.entdb_server.wal.sqs",
-    "dbaas.entdb_server.wal.servicebus",
-    "dbaas.entdb_server.wal.eventhubs",
-    "dbaas.entdb_server.tools.restore",
-    "dbaas.entdb_server.tools.schema_cli",
-    "dbaas.entdb_server.snapshot.snapshotter",
-    "dbaas.entdb_server.archive.archiver",
-    "dbaas.entdb_server.storage.s3",
-    "dbaas.entdb_server.storage.azure_blob",
-    "dbaas.entdb_server.storage.gcs",
+    "entdb_server.wal.kinesis",
+    "entdb_server.wal.kafka",
+    "entdb_server.wal.pubsub",
+    "entdb_server.wal.sqs",
+    "entdb_server.wal.servicebus",
+    "entdb_server.wal.eventhubs",
+    "entdb_server.tools.restore",
+    "entdb_server.tools.schema_cli",
+    "entdb_server.snapshot.snapshotter",
+    "entdb_server.archive.archiver",
+    "entdb_server.storage.s3",
+    "entdb_server.storage.azure_blob",
+    "entdb_server.storage.gcs",
 ]
 strict_optional = false
 warn_return_any = false
@@ -203,7 +111,8 @@ warn_return_any = false
 # =============================================================================
 [tool.pytest.ini_options]
 minversion = "7.0"
-testpaths = ["tests"]
+testpaths = ["tests/python"]
+pythonpath = ["server/python", "sdk/python"]
 python_files = ["test_*.py", "*_test.py", "bench_*.py"]
 python_functions = ["test_*", "bench_*"]
 asyncio_mode = "auto"
@@ -228,7 +137,7 @@ timeout = 60
 # Coverage
 # =============================================================================
 [tool.coverage.run]
-source = ["dbaas", "sdk"]
+source = ["server/python/entdb_server", "sdk/python/entdb_sdk"]
 branch = true
 omit = [
     "*/tests/*",
