+++ server/python/pyproject.toml
@@ -1,0 +1,84 @@
+[build-system]
+requires = ["setuptools>=68.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "entdb-server"
+version = "1.7.0"
+description = "EntDB server — tenant-sharded graph database with nodes and edges"
+readme = "../../README.md"
+license = "AGPL-3.0-only"
+requires-python = ">=3.10"
+authors = [
+    {name = "EntDB Team", email = "team@entdb.io"}
+]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Developers",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Topic :: Database",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+    "Typing :: Typed",
+]
+keywords = ["database", "graph", "tenant", "sharding", "event-sourcing"]
+
+dependencies = [
+    "aiokafka>=0.10.0",
+    "aiobotocore>=2.11.0",
+    "aiohttp>=3.9.0",
+    "grpcio>=1.60.0",
+    "grpcio-health-checking>=1.60.0",
+    "PyJWT>=2.8.0",
+    "cryptography>=41.0.0",
+    "PyYAML>=6.0.0",
+]
+
+[project.optional-dependencies]
+pubsub = [
+    "google-cloud-pubsub>=2.18.0",
+]
+servicebus = [
+    "azure-servicebus>=7.11.0",
+]
+eventhubs = [
+    "azure-eventhub>=5.11.0",
+]
+azure-storage = [
+    "azure-storage-blob>=12.19.0",
+]
+gcs = [
+    "google-cloud-storage>=2.14.0",
+]
+metrics = [
+    "prometheus_client>=0.21.0",
+]
+tracing = [
+    "opentelemetry-api>=1.20.0",
+    "opentelemetry-sdk>=1.20.0",
+    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
+]
+all = [
+    "entdb-server[metrics,tracing,pubsub,servicebus,eventhubs,azure-storage,gcs]",
+]
+
+[project.scripts]
+entdb = "entdb_server.main:main"
+entdb-restore = "entdb_server.tools.restore:main"
+entdb-schema = "entdb_server.tools.schema_cli:main"
+
+[project.urls]
+Homepage = "https://github.com/elloloop/tenant-shard-db"
+Documentation = "https://github.com/elloloop/tenant-shard-db#readme"
+Repository = "https://github.com/elloloop/tenant-shard-db"
+Issues = "https://github.com/elloloop/tenant-shard-db/issues"
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["entdb_server*"]
+
+[tool.setuptools.package-data]
+"*" = ["py.typed", "*.proto", "*.mmd"]
