+++ tests/python/integration/test_payload_roundtrip_python.py
@@ -33,10 +33,10 @@ from entdb_server.schema.types import FieldDef as ServerFieldDef
 from entdb_server.schema.types import FieldKind as ServerFieldKind
 from entdb_server.schema.types import NodeTypeDef as ServerNodeTypeDef
 from entdb_server.wal.memory import InMemoryWalStream
-from sdk.entdb_sdk import register_proto_schema
-from sdk.entdb_sdk.client import DbClient
-from sdk.entdb_sdk.registry import get_registry, reset_registry
-from tests._test_schemas import test_schema_pb2 as ts
+from entdb_sdk import register_proto_schema
+from entdb_sdk.client import DbClient
+from entdb_sdk.registry import get_registry, reset_registry
+from tests.python._test_schemas import test_schema_pb2 as ts
 
 
 def _free_port() -> int:
