+++ tests/python/unit/test_acl_v2_extended.py
@@ -23,7 +23,7 @@ import pytest
 
 from entdb_server.apply.acl import AclManager, Permission
 from entdb_server.apply.canonical_store import CanonicalStore
-from sdk.entdb_sdk.schema import AclDefaults, EdgeTypeDef, NodeTypeDef
+from entdb_sdk.schema import AclDefaults, EdgeTypeDef, NodeTypeDef
 
 TENANT = "test_extended"
 ALICE = "user:alice"
