+++ tests/python/unit/test_fts_sdk.py
@@ -14,7 +14,7 @@ import pytest
 
 from entdb_server.schema.registry import get_registry, reset_registry
 from entdb_server.schema.types import NodeTypeDef, field
-from sdk.entdb_sdk.client import DbClient
+from entdb_sdk.client import DbClient
 
 TYPE_PRODUCT = 201
